Solution to Categorize Students Based on Grades

+++ Categorize Students Based on Grades@@67877d61dc107486ad5c274e/code.c
@@ -10,10 +10,10 @@ int main(){
     struct Student student[num];
     for(int i=0;i<num;i++)scanf("%d %s &f",student[i].id,student[i].name,student[i].num);
     for(int i=0;i<num;i++){
-        if(student[i].num>=85){
+        if((int)student[i].num>=85){
             printf("Roll Number: %d, Name: %s, Grade: A",student[i].id,student[i].name);
         }
-        else if(student[i].num<=84 && student[i].num>=70){
+        else if((int)student[i].num<=84 && (int)student[i].num>=70){
             printf("Roll Number: %d, Name: %s, Grade: B",student[i].id,student[i].name);
         }
         // else {
